requirements.txt
fastapi==0.116.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
pydantic==2.11.7
pydantic_core==2.33.2
uvicorn==0.35.0
uvicorn-worker==0.3.0
